+++ Howest.Prog.Cia.CurrencySwapper.Core/CurrencyConverter.cs
@@ -6,7 +6,7 @@ namespace Howest.Prog.Cia.CurrencySwapper.Core
     {
         public const string AmountMustBePositive = "Conversion rate must be greater than zero";
 
-        public double Convert(double amount, double rate)
+        public decimal Convert(decimal amount, decimal rate)
         {
             if (rate <= 0)   //domain rule, koers kan nooit negatief of nul zijn
             {
